Remove: hooks 삭제
